patch

+++ www/controllers/Stat.php
@@ -52,7 +52,7 @@ class Stat
      *  It is possible to count the number of requests
      *  It is possible to add an offset to the request
      */
-    public function getRpmAccess(string $name, int $releasever, string $env, bool $count = false, bool $withOffset = false, int $offset = 0) : array|int
+    public function getRpmAccess(string $name, float $releasever, string $env, bool $count = false, bool $withOffset = false, int $offset = 0) : array|int
     {
         return $this->model->getRpmAccess($name, $releasever, $env, $count, $withOffset, $offset);
     }
